refactoring - HomeScene MVVM 패턴 적용 중
