+++ STUDYHUB2/Assets.xcassets/LogoImage.imageset/Contents.json
@@ -1,19 +1,21 @@
 {
   "images" : [
     {
+      "filename" : "logo.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
+      "filename" : "logo 1.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "logo 2.png",
       "idiom" : "universal",
       "scale" : "3x"
     },
     {
-      "filename" : "0540f2a0d7629f25.png",
       "idiom" : "iphone",
       "scale" : "1x"
     },
