+++ STUDYHUB2/View/MainScene/DetailsViewController.swift
